requirements.txt
numpy
scipy
torch
torchvision
pytorch-lightning
pillow
opencv-python
openmim
anime-face-detector